[ui] Add script to generate globals.css from theme.ts (#1239)
Summary: Create script that generates globals.css from a template and
theme.ts. Makes it so we have 1 source of truth for the colors.

Relevant Issues: Closes GML-253

Type of change: /kind cleanup

Test Plan: Ensure that global.css still works in the dev ui, try
changing a value in theme.ts that is used everywhere and see it change
in the UI.

Ensure that the workflow actually picks up the changes as well.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: ecfbbf7d1938c54aacf27c4cd9e6568c29d8d92b