Fix jetstream consumer initialization and metricsprocessor acking (#1319)
Summary: We were incorrectly cleaning up our partitions. This fixes that
behavior for now, but we should eventually move to using NATS's Nack for
managing consumers and streams.
This also updates metrics processor to make sure we are actually
acking/nacking our messages in all cases.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: check that metrics are streaming to metrics processor with a
skaffold redeploy

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: ddc2e5d95f44ce6128204a059b3867c29b40be3d