Support partitioning into non powers of 2 (#1329)
Summary: The old partitioning math only supported perfect divisors
of 16^3 (i.e. powers of 2) for number of partitions. This change
updates the math to ensure that we don't round the fraction too
early and also always include remainders in the last partition.

Partitions should probably still be powers of 2 to have even
distribution
but this just ensures that other settings don't cause us to miss any
messages.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Added a test with a parition count that is not a power
of 2. Saw failures and lack of `fff` inclusion before this change.
Now the paritions make sense.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 8c0d2846ee2098a0505637bd5c3efcdf0bda5865