Add the docker commands to the install script (#1331)
Summary: This adds an oci_push rule for the gem. Updates the install
script to add in the device detection, docker mounts for host libs,
CUDA etc. Also adds a mount for a dir to write model files into.

Relevant Issues: Towards GML-166

Type of change: /kind feature

Test Plan: Copied the script over to a Jetson device and ran it.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: e966143d7b0a129097de148b803a9935176c5878