[build] Build openh264 statically (#1340)
Summary: Building openh264 as a shared library led to some strange
behaviour with exception handling. The libc++abi function definitions
would end up undefined in the main library and only defined in the
openh264 shared lib. For unknown reasons, this led to crashes when
exceptions are thrown (even caught exceptions). To bypass this issue we
build openh264 as a static lib instead of a shared library.

Relevant Issues: Towards GML-148.

Type of change: /kind cleanup

Test Plan: Tested that throwing an exception with openh264 linked in, no
longer causes a crash.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d2c10ea1e72fa3ab9c56a48121ee7fca3ef20017