[gem] Add ModelBuilder for openvino models. (#1339)
Summary: Adds a `ModelBuilder` for creating openvino models from onnx
files. Also adds a default execution context builder for the openvino
execution context. I haven't found an openvino API for loading an onnx
model from an in memory blob, so I add to (at least temporarily) modify
the `BlobStore` interface to support returning a file path to a blob. I
think the storage interfaces are likely to change anyways when we
download blobs from the controlplane instead of relying on them existing
on the filesystem.

Relevant Issues: Towards GML-274.

Type of change: /kind feature.

Test Plan: Added a simple test for the model builder that ensures it
doesn't error when building a toy model.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: c6c4945f1aabceab87a3a3f6fa9f2e5f88b125c9