[gem] Add OpenVinoExecuteCalculator to execute models with openvino. (#1341)
Summary: Adds a calculator to execute a compiled openvino model. The
calculator currently expects all of the input tensor shapes to be given
explicitly as options to the calculator.

Relevant Issues: Towards GML-274.

Type of change: /kind feature.

Test Plan: Tested by running the full gem binary on auger with some
other changes. I will add a unit test for this calculator in the future.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 897e8248a3123bba7355a24781b36033b9715dd6