Add protos for fetching executionGraphs from a pipeline (#1302)
Summary: This adds protos for
- Sending a request to the fleetmgr service to deploy a model pipeline
to a specific device.
- Sending a request to the model mgmt service to get the model execution
graph for a pipeline, given a pipeline ID.
Note in this design, the conversion from pipeline to model execution
graph is done in the control plane.
This adds the DAG proto message to typespb, since this is a generic
message type which can be used for other DAG types (for example,
pipelines).

Relevant Issues: fixes GML-258

Type of change: /kind feature

Test Plan: this just adds protos but doesn't use them. existing tests
should pass.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 2d9352b656ad652392c3ec6c8490a8e063a19063