Add protos for managing pipeline lifecycle between cp/edge (#1303)
Summary: This PR adds the protos for managing a pipeline execution on an
edge device. The CP sends out the ApplyExecutionGraph message with its
desired spec. The GEM is responsible for reconciling its state with the
spec, and sending out updates to the CP accordingly.

Relevant Issues: fixes GML-260

Type of change: /kind feature

Test Plan: none yet, these are new protos that are unused

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 62719eba1fc159da406161f6e674d8377d36b44c