[sysroots] Add mesa libs to runtime sysroot (#1371)
Summary: We need the mesa libs at runtime, so this PR includes them in
the runtime sysroot. We could attempt to load them from the host but
this could be a bit brittle especially if the host doesn't have them
installed.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Built an image, saw that the expected mesa shared libraries
are now there.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 256fe5d355871fdbc1d4d2c3e184119f3a70a3fe