[gem] Support passing custom /sys/class/net path to GEM binary (#1372)
Summary: Adds a command line flag to the GEM binary to specify where to
look for /sys/class/net devices. This allows the host devices to be
mounted into a container at a custom path.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested by deploying a container with /sys mounted at
/host/sys and passing `--sys_class_net_path=/host/sys/class/net` and saw
that it still finds the same MAC address.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 4d521446b260e05579ea1feb0835e9768a78f09b