Add stream for device connected (#1376)
Summary: We want to know when a device has connected. For example, when
a new device has connected, we need to find all pipelines that should be
deployed to the device and send the deploy messages. This PR adds the
stream for the device connection messages.

Relevant Issues: towards GML-265

Type of change: /kind feature

Test Plan: unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 9f559ada793bf4cd85b682efa15ee89d24d85853