[docs] Add instructions to setup namespace in DEVELOPMENT.md (#1409)
Summary: Update docs to set the namespace for kubectl that is pretty
useful.

I switched to auger and this is something that I liked in my old env
given our dev namespace setup.

Relevant Issues: n/a

Type of change: /kind cleanup

Test Plan: Ran as part of my setup

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: ff2a13ad9c269a1d30941b811921ed2149485955