[controlplane] Define and implement GetAugmentedTokenForDeployKey on auth service (#1410)
Summary: Part 1 of switching over to using a single endpoint for envoy
auth

Adding a DeployKey AugmentedToken endpoint that generates bearer for
deploy-key clients. Basically moves the contained header logic from
GetAugmentedToken into a separate rpc. We'll call this rpc inside of the
AugmentedAuthHandler of the api service. That change is in a later diff.

Relevant Issues: Towards GML-275

Type of change: /kind cleanup

Test Plan: Added tests and integrated this flow into a hacky branch that
worked.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: b95b77fac7074649e3c287724a2497c06520ec85