Update durablepartitionHandler to accept CPNATSTopics (#1399)
Summary: Previously, the durablePartitionHandler assumes that you are
listening to messages on EdgeCPNATS topics. However, we are also adding
control-plane specific messages that we want to use durableChannels for.
This PR updates the durablePartitionHandler so that it can take both
cases.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 6cb372008e16fb6292daeda829e4f47a348fb3d8