Add Device Tag APIs to FleetEdgeMgr (#1395)
Summary: This adds the following 3 RPC calls to FleetEdgeMgr

- `GetTags`
- `UpsertTag`
- `DeleteTag`

This API works with unordered tags. Any ordering is the responsibility
of the UI.
To help the UI sort tags, `GetTags` returns a `TagWithMetadata` which
contains an `UpdatedAt` timestamp.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Add unit tests.

---------

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: b65c731b45dca478fd400e82a33d202707bc1ef4