ui: add linter for visspec schema (#1455)
Summary: The example visspec json fell out of date with the schema. This
PR corrects the issue, and also adds a linter to prevent these
regressions in the future.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: ran it with and without a schema error on the example file

GitOrigin-RevId: 3d4f14b6daca11866b5a90174bd0a33831d6272a