Add GetFleetByName proto (#1540)
Summary: We have a bug with our current GetFleetByName, where it may
find a fleet in another org. Rather than adding an additional orgID into
the GetFleet request, this PR just makes a new proto for it.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: none yet, this only adds the proto but doesnt use it

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: f2d98fcf445c57e605475521d2dbea1e1bd7d729