Fix typos in Go code (#1552)
Summary: Found some typos with codespell in our Go code after I noticed
a bunch in the UI code. I don't think it makes sense to do codespell as
a default linter, since it can be noisy, but we can consider it.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: N/A

GitOrigin-RevId: 58dad2c8a3a516203cd1d244607e077fe64d5294