Fix typos in C++ parts of code (#1553)
Summary: Found some typos with codespell in our C++ code after I noticed
a bunch in the UI code. I don't think it makes sense to do codespell as
a default linter, since it can be noisy, but we can consider it.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: N/A

GitOrigin-RevId: 8265c7d883221573d8c8ce1d88020033e2319455