[gem] Handle ApplyExecutionGraph messages and use it to get the model spec (#1548)
Summary: This updates the model exec handler to read the provided
message, unpack it to get the model and run the provided spec
instead of the hardcoded spec.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Ran a gem on the jetson and verified that it works e2e.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: cb51126abe7478a362080906684f5b4d3768cdb6