[jetstream] Reduce consumer replicas to make NATS stable (#1587)
Summary: NATS tips over with the addition of all the new durable
Jetstream consumers. Part of the problem is that we run with 3 replicas
on the consumers and 4096 consumers per Stream that wants to listen to a
Jetstream Stream.
A quick fix is to move the number of consumer replicas from 3 -> 1. It
reduces availability of consumers for now, but because of stream
replication, we can still resume Consumers when a NATS node returns.
Susceptible to file corruption and disk failure so we'll need to figure
out some more changes we want.

This change reduces the measured RAM usage by ~6x on my cluster with 0
devices (3.7GB to 650MB) and NATS calls continue to work.

Relevant Issues: n/a

Type of change: /kind cleanup

Test Plan: Tested against skaffold and ensured that I could access
Jetstream.

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 6d2e3469ae18ba3dc69d6763537bf284012e4228