[nats] Add missing durable stream EDGE_CP_TOPIC_EXEC to durables list (#1616)
Summary: We assume this stream is durable elsewhere, just not in the
streams list. Without this any such messages that arrive from the Device
will never meet the end consumer because we don't prepend "Durable" to
the topic.

Relevant Issues: n/a

Type of change: /kind cleanup

Test Plan: n/a

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 43c3f4133c3ec0f4e7be1e3517c3d73e1ad8cc10