[apt_parse] Add spec for apt package requirements (#1640)
Summary: This is the beginning of adding the `apt_parse` tool. The tool
is similar in concept to `pip_parse` for python deps but for apt deps.
The `apt_parse` tool will eventually take in a set of specs for apt
dependencies in yaml form (think `requirements.in` equivalent in
`pip_parse`) and output a `.bzl` file containing repository rules for
each transitive dependency of the explicit dependencies listed in the
spec. The `apt_parse` tool takes some pieces from the old
`package_satisfier`, however, by simplifying its scope it should be much
cleaner than the `package_satisfier`.

This PR adds the specifications for apt parse dependencies and the spec
for a "Pinned" apt dependency. Each set of apt dependencies can list a
set of debian-style repository indices to search for packages and their
dependency graph information. For example, we could have a set of
dependencies from the `debian12` repos, and another set from NVIDIA's
jetson repos.

Relevant Issues: Towards GML-251.

Type of change: /kind infra

Test Plan: No tests for just the spec yet. Other pieces of the
`apt_parse` tool will have tests.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d0a7ab7190cc6dfb0070c0ae9d99befc3ab24dcc