[apt_parse] Add debian index parser (#1642)
Summary: Adds a parser for debian-style index files. The parser builds a
set of `PackageOptions` for each package name. Each option can be a
different version of the package, or package that provides the virtual
package with the given name.

Relevant Issues: Towards GML-251.

Type of change: /kind infra

Test Plan: Added simple unit test for the parsing behaviour.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: cdbb71fa7c8642b939e0d46854e903b291a892f6