[FleetMgr] Add device capabilities to DeviceInfo (#1657)
Summary: Add capabilities to DeviceInfo proto. We want to do this so we
don't have to restart the physical server with a new env flag anytime we
register a device with a different runtime.

In the future, will wire up the other RPCs to read/write this data
from/to the DB.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: n/a proto change.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: 0e415476b6057c179059ed69b6e0a9c5a64b5897