[nats] Don't block nats subs on errors (#1634)
Summary: We used to NoAck messages on handle message error but that
would just burn CPU if we ever had an error on the handler. For now we
will just drop the message and write an error. In the future
([GML-286](https://linear.app/gimletlabs/issue/GML-286/jetstream-message-handler-failures-should-send-a-sentry-message))
we should send a sentry message.

Relevant Issues: n/a

Type of change: /kind feature

Test Plan: Added a test.

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
Co-authored-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 7fb43d6e2550534849d3c97f695de8e7f1b13383