[toolchains] Split clang toolchains into two repos (#1656)
Summary: Splits the `clang_toolchain` bazel repos into two bazel repos
each. One repo contains the implementation of the clang toolchain, and
the other repo contains the toolchain definition. This allows all of the
toolchain definitions to be registered with bazel without pulling in the
dependencies for those toolchains. Then when bazel resolves the
toolchain to use for a given config it will only load the implementation
repo for the resolved toolchain.

Relevant Issues: Towards GML-251

Type of change: /kind infra.

Test Plan: Built everything with both `x86_64_sysroot` and
`jetson_sysroot`.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 6deebfbc4b7317dfe1f146caf0b9281df3c00d9b