[sysroots] Add bazel macro to create new style sysroots (#1662)
Summary: Adds a bazel macro that creates a sysroot from a list of
targets that provide `rules_pkg` providers. The macro creates a bazel
repo that symlinks all of the files from the `rules_pkg` providers, and
also uses `rules_pkg` to build a tar file for the sysroot.

Relevant Issues: Towards GML-251.

Type of change: /kind infra.

Test Plan: Tested end to end with the rest of the sysroot changes.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: af15271e48b1523432ace43e8369674a51a488b6