[pg] Switch rest of the pg calls to Contexted calls (#1708)
Summary: We want to use contexts in all pg calls so they can be
cancelled if a context is cancelled. This ensures they are added.
The only thing missing was tx.NamedQuery, but I made sure that
transactions were all context-ed so they should be cancelled if
something happens.

Relevant Issues: n/a

Type of change: /kind cleanup

Test Plan: Ensure tests run

---------

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: ad96e835853a364bc544d7d0f7a362f58780006e