[Pipeline Reconciliation] Newly connected devices should always receive pipeline specs (#1625)
Summary: We will be updating the GEM to manage multiple pipelines/send
out status updates regarding how the pipeline is running. In the current
state, we correctly send out the pipeline to the GEM when it starts up.
This is because the GEMs do not send statuses yet. When a GEM starts up,
we want to send the execution graph regardless of the last reported
status.
The alternative to this is to send a "GetPipelines" request from the GEM
to the control plane when it starts up. We can also get the same effect
using the DeviceConnected message, without having to create a specific
`GetPipelinesRequest` and `GetPipelinesResponse`.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: unit tests

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: d51215f9c46907184e346df8ef9a5f157f541f5c