[gem] Fix test images that pulled in sysroot twice. (#1719)
Summary: The sysroot was being pulled in twice into some of our manual
test image targets. The second time was as runfiles to the test binary
because of a dependency on the test runner. Since these images don't run
the test runner, this dep was unnecessary. This PR removes the
dependency, and as a result removes build issues trying to pull the
sysroot in as runfiles.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran `test_on_jetson.sh` for one of the argus nvbufsurf tests.
Followed its instructions and the test passed.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: d4a8ea5e0499a6c41b1908a414d963f95b95326a