[sysroots] Add sysroot for intel gpus. (#1710)
Summary: Adds a sysroot that includes the necessary libs to use an intel
GPU. Also, cleansup a bug around `ManualPackages` in the `apt_parse`
tool.

Relevant Issues: Towards GML-276.

Type of change: /kind feature

Test Plan: Tested along with another PR, that I can deploy the GEM to a
device with an intel GPU and have it run the model using the GPU.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 0a8337bd477c73f249c1b4ca83da4aa8c3a56f56