[gem] Support building GEM with `intelgpu_sysroot`. (#1712)
Summary: Add support to build a Intel GPU version of the GEM. This adds
the bazel config `intelgpu_sysroot`. Also, it changes the `install.sh`
script to detect Intel GPUs and then run with the intel GPU image.

Relevant Issues: Towards GML-276.

Type of change: /kind feature

Test Plan: Tested end to end on a nuc. Saw that the model was running
using the GPU.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 07d7ff096991cacfecc2c590e6b22187c539abe0