[pipeline] Model execution sends status updates to controlplane (#1636)
Summary: This PR does two things:
- Modify the model execution handler so that it can track/manage
multiple pipelines at once.
- Modify the model execution handler so that it sends status updates to
the control plane, such as when it is ready/terminating.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: run GEM, deploy model to it. I don't have a working jetson,
so just verify that the "terminating" status is set in the DB once the
GEM starts up.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 480c8fb17aaa425842511afc3098ed46addceca8