[gem] Remove MapReadOnly from BlobStore interface (#1747)
Summary: Removes `MapReadOnly` from the `BlobStore` interface so that
there's only one path to get a blob. We might want to consider renaming
the interface to `FileStore`. Since the TensorRT engine file can't be
loaded by path, I moved the mmaping utils to `src/common/system`. This
also removes a bug that was in `CachedBlobStore`'s `MapReadOnly`.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested by deploying to a jetson. Saw that it downloads files
as expected. Also tested with another change, to ensure that it loads
the engine file correctly.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 5f18d95dad268f86d3ae4f0b47930421944c7965