[CachedBlobStore] Use `by-key` directory if file not found (#1748)
Summary: Try to read files from the `by-key` directory if they can't be
found in the `CachedBlobStore` map. This allows things that were
previously upserted to be reused after device restart.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested on a jetson. Saw that the engine was loaded from the
by-key directory.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 57918da05bc60fb3144757b060edc11e1a885ccc