Relax buf breaking change to only check for WIRE compatibility issues (#1791)
Summary: We allow fields to be deleted and reserved and types
to change as long as the wire format is compatible. The `FILE` level
checker is more comprehensive than we like and instead we should use
the `WIRE` level checkers to line up with our dev practices.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: The linter no longer fails if a field is deleted and it's ID
is reserved.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: f44282d1d513ce8ac486ebf07495ca9c22e04828