Add cookbook for deploying a node to our nuc cluster (#1771)
Summary: This PR adds a cookbook which does the necessary setup on the
NUCs to run the NUC as a K3s agent. It expects the k3s credentials
(token, host IP) to be added to `templates/credentials`.

Relevant Issues: towards GML-291

Type of change: /kind feature

Test Plan: deploy to some NUCs, verify that they have connected to the
k3s cluster and can run a pod.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: d13d569c5aa70731237c0b8b88d0c9ea89f83cf5