Add flag to allow changing log levels for services (#1809)
Summary: It's convenient to be able to change the logging level on
a svc without needing to rebuild it. So expose a CLI flag to do the
same. Also added a log that I found useful to check if metrics were
being processed.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: skaffolded to dev cluster.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: aac2c2487bcd1f0c6a0c619bee6ebec3e4dd7b0c