Update VideoHandler to start/stop and keepalive as needed (#1845)
Summary: The video handler used to ignore the inbound message info
from cloud and always send video data. Now instead it respects
Start and Stop. It also includes a 5 second timer for KeepAlive messages
after which it automatically calls stop if no such messags are recieved.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: skaffolded a gem and then checked the logs when watching
the video and closing the tab.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: d4f50288977c7fdcb6fbc8f2e155682af1ef66c1