[gem controller] Split Init() into Register() and expose getters() for private vars (#1851)
Summary: Expose fields and add Register() call to the controller. This
is in prep for the fake gem which uses these function calls as a child
of the Controller class.

Relevant Issues: Towards GML-297

Type of change: /kind feature

Test Plan: Ensure that the gem still sends and receives data with the
controlplane

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>
GitOrigin-RevId: b3ea3f14adc9b2ccf76f5e56af3c2b75c4655f72