Use timestamps from opencv in cam calculator (#1849)
Summary: I am hoping to use the OpenCV Calculator to open
pre-recorded video files in addition to camera devics. However
we need a way to backpressure and control the framerate on video files.

We were using the frame count as timestamps. Instead this
switches over to using timestamps from opencv so that we can use
PacketThinnerCalculators and PacketResamplerCalculators to try to
adjust the framerate.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran a gem binary, made sure that the opencv cam worked
and the model ran as expected. Also ran the manual test and inspected
the single frame it generates.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 8d59bf96ec551ef44adf9c24ff1e04d80b7c897c