Remove xvfb: no longer needed with Argus change (#1653)
Summary: Discovered that the only reason we needed a fake X server is
because DISPLAY was set. But we can work around this by unsetting that
environment variable. This means we don't need the fake X server
anymore, so remove it.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Build and deploy GEM with the changes.

Signed-off-by: Omid Azizi <oazizi@gimletlabs.ai>
Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
Co-authored-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: f65d74cc1662740eecf5228eb657e8a00727fb0e