Ensure stack protection size is a multiple of page size (#1870)
Summary: The min stack size from the kernel is not always a multiple of
the page size. The original code was working as long as the stack size
was < 4 * page size. This ensures we always
pick the guard size to be a multiple of the page size.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Run on hammer which has a lager stack protection size.

Signed-off-by: Zain Asgar <zasgar@gimletlabs.ai>
GitOrigin-RevId: e0db495b3cbde71b36a3e4f0177a1e1d36a8ec32