Support additional symlinks is sysroot generation (#1867)
Summary: Add symlink param to handle packages that rely on
update-alternatives to setup symlinks.
This ensures that `/etc/rmt` resolves in the sysroot and doesn't
cause failures.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Ran a build and test with the x86_64_sysroot and saw the BES
upload logs go away. Checked my sandbox and ensured that `/etc/rmt`
resolves to `/usr/sbin/rmt-tar` via `/usr/sbin/rmt` as expected.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 686fde95a69383559e39b0e6fbddd4a379f4badf