[openvino] Patch openvino to avoid issues with int8 models (#1904)
Summary: There seems to be a missing logical not, in OpenVINO's compiler
passes. This causes int8 ops to not get converted correctly. In
addition, there's an issue with the GPU plugin that causes int8 models
not to run due to compression to fp16 before running the int8 compiler
passes. This PR addresses these two issues with small patches to
OpenVINO.

Relevant Issues: Towards GML-340.

Type of change: /kind cleanup

Test Plan: Tested a quantized model on CPU and GPU. Used the debug
capabilities to dump graphs for the models and saw that they are in fact
running the quantized version (on both CPU and GPU).

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 16593ec412df1a2820f44739e6ea80b0e10f13d9