[FleetEdgeMgr] Add createdAt to tag metadata (#1899)
Summary: We decided that we want arrays of tags returned by GQL to
be returned sorted by their creation timestamp. We considered using
the existing `updatedAt` tag metadata for sorting, but it's not as
stable as `createdAt`.

This PR updates the `FleetEdgeMgr` to add the `createdAt` metadata
in the tags it returns.

I will update the `FleetMgr` to do the same in #1850.

Relevant Issues: towards GML-328

Type of change: /kind feature

Test Plan: Update the unit tests.

---------

Signed-off-by: Hannah Troisi <hannah@gimletlabs.ai>
GitOrigin-RevId: 9a6c504e17b7a4baef2720557d1f337217de2cfc