[pipeline execution] GEM pipeline execution should update and replace running pipeline (#1905)
Summary: We want to change our GEM to run a single pipeline at a time.
In the future, we will actually merge multiple pipelines to run on the
GEM. For now, if we are already running a pipeline and a new pipeline is
executed, we should stop the currently running pipeline and start the
new one.

Relevant Issues: towards GML-352

Type of change: /kind feature

Test Plan: run GEM, deploy a pipeline, then deploy another pipeline.
observe that the first pipeline stops, and the second one starts.

Signed-off-by: Michelle Nguyen <michelle@gimletlabs.ai>
GitOrigin-RevId: 0f80ffd2fb3ae1539bff12264f081bc7d65df76b