[apt_parse] Mirror deb packages to GCS. (#1918)
Summary: The snapshot debian repositories are very slow. This mirrors
the packages to GCS so that we don't rely on the snapshot repositories.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Tested that the mirrors upload correctly, and that I can use
the sysroots from the mirrors.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: c44fed6e0183f4cbb89a1a0305dd511f572ec3d7