Support video files in OpenCV cam (#1931)
Summary: This adds a param that allows one to override the reported
devices from OpenCV. The expectation is that the overriden path is
a video file that will be used instead of the camera.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Ran the gem binary locally with a video file. Checked that
the
model ran.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 1c266a6171f0f6f10135e3139d53f7b10c3a3b73