Update dev fake camera setup to not need k8s device plugin (#1947)
Summary: The device plugin made it somewhat challenging to use
arbitrary video files. The setup was also tied to using a particular
cluster. This setup instead should work on other clusters (as long as
the initContainer has acls to pull from the given GCS location).

This also makes it possible for people to pick which video they want to
use.

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Ran the skaffold command on the dev cluster with various
video files.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 73082b0673c7e03921ee2b279b34ac4529286d4c