Drop unused fake_camera plugins and setup (#1949)
Summary: The fake camera setup on the k8s dev cluster was challenging
to use. We can now directly pass in a video file to the GEM so this
can be cleaned up. Note that it may still be useful to proxy a video
through a fake camera device for testing. Devs can manually run
`modprobe v4l2loopback` to create such a device and use it if so
desired.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran terraform apply

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: cdf9b7aa9046881409d88d4a644291d4ac353a94