Fix argus default video graph (#1951)
Summary: There were a couple of bugs in the default video graph for
the Argus/Jetson device. This fixes them.
Namely the SubgraphOptions proto type was wrong. Additionally,
the output from the camera was using a tagged output instead of
an index based output.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Deploying a GEM on jetson shows the default video now.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 981843f286aabfa3605153d6c1c33c2c5e23de88