Add video_header to our mediapipe camera sources (#1953)
Summary: Adding the video_header packet allows the encoders to get
more information about the stream and encode the video better. This
also drops the frame_rate packet which will no longer be used for
encoding (except in tests).

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Ran the GEM locally and on a Jetson.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: fe1d4f945af1bf5048ccddc99342ab4dc7a85551