[jetson] Fix EGL_EXT_client_extensions not supported error (#1979)
Summary: NVIDIA's container runtime toolkit, uses `ldconfig` to update
the `ld.so.cache` on container startup. It seems that this process is
non-deterministic, since sometimes the EGL ICD loader is able to find
`/usr/lib/aarch64-linux-gnu/tegra-egl/libEGL_nvidia.so`and sometimes its
not. When its not able to find that library, it defaults to the
vendor-neutral implementation which doesn't seem to work on the NVIDIA
hardware leading to the error people have commonly seen:
`EGL_EXT_client_extensions not supported`. This PR fixes the issue by
ensuring the nvidia implementation lib is in the `LD_LIBRARY_PATH`.

Relevant Issues: N/A

Type of change: /kind bug.

Test Plan: Tested by running a binary in our container, that just
queries for the relevant extension support. Without the fix, it fails
about 50% of the time. Ran it 100 times with the fix and didn't see any
failures.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 6739ef3a50e76398cce8bf96043b947268d01edd