Rescale video while maintaining aspect ratio using ScaleImageCalculator (#1955)
Summary: This rescales the video using ScaleImageCalculator.
We should pass in the contain size via graph/subgraph options,
but they are currently hardcoded. (640x480 for the video size).

Any letterboxing or pillarboxing should be done by the UI if we
so desire.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Ran a dev gem with inputs of various sizes. Ensured that
the output video was always scaled and smaller than or equal to
640x480

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: fe8f714eef67f0fdcb1a153fad77786e032a51b2