Update ImageOverlayCallback to accept arbitrary protos (#1996)
Summary: In preparation for being able to send Video Header metadata
to to the VideoSink, this updates the VideoWithOverlaysCallback to take
a vector of abitrary proto messages and just forwards them all on the
bridge.
Some work needed to be done to set the frame_ts and eof properly in this
new setup.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ensured that the unit test works. Also tested end to end with
a GEM.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 28412aa52d03ca9dad98f4d547a614a971f8430b