Update mediapipe proto (#2005)
Summary: This adds the VideoHeader. Also updates the TS definitions
to include the JSON `@type` attr and updates the types to line up
with the JSON types (so int64s are strings).

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Ran the devUI and made sure it still works.

GitOrigin-RevId: aa62221bf8ad3bcfd642523c487ce006ceb64e34