Propagate VideoHeader to the UI (#2003)
Summary: This propagates the VIDEO_HEADER to the Overlay calculator
which then sends it to the UI. We also update the UI to use the
video dimensions from the header to render an appropriately sized
canvas. This means we no longer rely on the inferred dimensions
from the Broadway player.

Relevant Issues: Fixes GML-335

Type of change: /kind feature

Test Plan: Ran the dev UI with a GEM with these changes.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: d293084f78e645369a530a578d1cd28a94c21d99