Replace rules_meta with with_cfg (#2006)
Summary: rules_meta relies on the name parameter for rules which
is removed in bazel7. So this switches us over to it's replacement
which is with_cfg.bzl
For more info see fmeum/rules_meta#1

Relevant Issues: N/A

Type of change: /kind infra

Test Plan: Multiarch images still build.

Signed-off-by: Vihang Mehta <vihang@gimletlabs.ai>
GitOrigin-RevId: 742d9f8479883ec9ca224f4565dd3571d993e530