Make FlowLimiter wait for all outputs (#2029)
Summary: We observed multi-second latencies when we introduced the
ImageQuality calculations to the mediapipe graphs. In the case of the
default graph where no model is running, (before this PR) the flow
limiter allowed frames as soon as the video encoder was finished. This
causes an unbounded queue of frames in the ImageQualityCalculator's
input stream, because the ImageQualityCalculator cannot keep up with
realtime. Since the Overlay sink waits for the image quality results,
this led to unbounded increases in latency. This PR fixes the issue by
adding an output signal on the overlay sink, which can be used to signal
the flow limiter to allow frames. This ensures that the queue size is 1
for all calculators whose output data path ends up in the overlay sink.

Relevant Issues: N/A

Type of change: /kind bug.

Test Plan: Tested with both the default graph and the barcode graph,
that there's no longer long latencies.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 15eec9191ebc001e5218a7d3caa7c7669cb75553