[gem] Handle TimestampBounds correctly (#2100)
Summary: We need to set `TimestampOffset` to zero in order to let
mediapipe know that any timestamp bounds can be propagated through our
calculators. In addition, we need to make our overlayed video sink
always run `Process` even if its only given timestamp bounds. This
allows it to output video even if it receives timestamp bounds for the
detection input. This change fixes issues when loop calculators are
introduced in order to classify each detection separately.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Tested that a set of `Begin/EndLoopCalculators` will no
longer cause the system to hang when the vector its looping over is
empty.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 0c46e445264d43f23f1c94bc3df9fab6d24ea1c5