[gem] Add ScoresToLabelCalculator for adding classification labels to detections (#2099)
Summary: Adds a calculator to convert scores into class labels. This is
a bit of a hack currently as it just modifies an existing Detection
proto. We should think through the API we want for classifiers on top of
detectors, and once we have that design more sorted out we can modify
this calculator.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Tested by deploying to a NUC with both a product detector and
an embedding based product classifier.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 8d5cbdbec0ee3dd5f0737d2618cabbf4e058bd95