[gem] DetectionsToMediapipeCalculator support single proto as well as vector (#2102)
Summary: This calculator will be used as part of the classification
pipeline. To better support that use case this PR adds support for
single detections, in addition to vectors of detections.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Updated the test, also tested end to end as part of the
classification pipeline.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 9ccefdeeffbdbe972b34519376cd05bd54252b49