[gem] Output timestamp bound instead of empty detections vector. (#2106)
Summary: Outputs a timestamp bound instead of a an empty vector of
detections when the model outputs 0 detections. This works better when
using mediapipe's loop vector calculators.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Tested that the video still gets output when there's no
detections. Also, tested with the vector loop calculators to see that
they work with the timestamp bound.

Signed-off-by: James Bartlett <james@gimletlabs.ai>
GitOrigin-RevId: 5e51a3ae4672cf4d05de0b0aa15ce112da7ebee8